Move PipelineContainer back into 'pipeline'
