+++ source/gloperate/CMakeLists.txt
@@ -45,7 +45,6 @@ set(headers
     ${include_path}/base/ResourceManager.inl
     ${include_path}/base/AbstractCanvas.h
     ${include_path}/base/Canvas.h
-    ${include_path}/base/PipelineContainer.h
     ${include_path}/base/AbstractContext.h
     ${include_path}/base/AbstractGLContext.h
     ${include_path}/base/AbstractGLContextFactory.h
@@ -67,6 +66,7 @@ set(headers
     ${include_path}/base/Storer.inl
     ${include_path}/base/GlmProperties.h
 
+    ${include_path}/pipeline/PipelineContainer.h
     ${include_path}/pipeline/Stage.h
     ${include_path}/pipeline/Stage.inl
     ${include_path}/pipeline/Pipeline.h
@@ -125,7 +125,6 @@ set(sources
     ${source_path}/base/ResourceManager.cpp
     ${source_path}/base/AbstractCanvas.cpp
     ${source_path}/base/Canvas.cpp
-    ${source_path}/base/PipelineContainer.cpp
     ${source_path}/base/AbstractContext.cpp
     ${source_path}/base/AbstractGLContext.cpp
     ${source_path}/base/AbstractGLContextFactory.cpp
@@ -140,6 +139,7 @@ set(sources
     ${source_path}/base/AbstractLoader.cpp
     ${source_path}/base/AbstractStorer.cpp
 
+    ${source_path}/pipeline/PipelineContainer.cpp
     ${source_path}/pipeline/Stage.cpp
     ${source_path}/pipeline/Pipeline.cpp
     ${source_path}/pipeline/AbstractSlot.cpp
