+++ source/gloperate/include/gloperate/pipeline/PipelineContainer.h
